test(onboarding): real-pg atomic promotion on approval (membership + users projection)
